# src/Facilities/Domain/Values/SeatNumber.php
<?php

namespace App\Facilities\Domain\Values;

use App\Core\Domain\IntegerValue;

readonly class SeatNumber extends IntegerValue
{
    protected static function validate(int $value): void
    {
        if ($value < 1) {
            throw new \InvalidArgumentException('Seat number must be at least 1.');
        }
    }
}